機能面を統合
