+++ src/components/Main/MainView/QallView/ScreenShareComponent.vue
@@ -5,13 +5,14 @@ import type { TrackInfo } from '/@/composables/qall/useLiveKitSDK'
 import AudioTrack from './AudioTrack.vue'
 import VideoTrack from './VideoTrack.vue'
 
-const { trackInfo, audioTrackInfo, participantIdentity } = defineProps<{
+const { trackInfo, audioTrackInfo } = defineProps<{
   trackInfo: TrackInfo
   audioTrackInfo: TrackInfo | undefined
-  participantIdentity: string
 }>()
 
 const { removeVideoTrack } = useQall()
+
+const volume = ref(1)
 
 const videoElement = useTemplateRef<HTMLVideoElement>('videoElement')
 
@@ -29,15 +30,23 @@ onUnmounted(() => {
 </script>
 
 <template>
-  <div :id="'camera-' + participantIdentity">
+  <div :id="'camera-' + trackInfo.participantIdentity">
     <div>
-      <p>{{ participantIdentity }}</p>
+      <p>{{ trackInfo.participantIdentity }}</p>
     </div>
     <VideoTrack v-if="trackInfo.trackPublication" :track-info="trackInfo" />
     <AudioTrack
       v-if="audioTrackInfo?.trackPublication"
       :track-info="audioTrackInfo"
+      :volume="volume"
     />
+    <input v-model="volume" type="range" min="0" max="1" step="0.01" />
+    <button
+      v-if="!trackInfo.isRemote && trackInfo.trackPublication"
+      @click="removeVideoTrack(trackInfo.trackPublication)"
+    >
+      Remove Screen Share
+    </button>
   </div>
 </template>
