+++ src/components/Main/MainView/QallView/VideoTrack.vue
@@ -1,13 +1,10 @@
 <script setup lang="ts">
 import { onMounted, onUnmounted, ref, useTemplateRef, watchEffect } from 'vue'
-import { useQall } from '/@/composables/qall/useQall'
 import type { TrackInfo } from '/@/composables/qall/useLiveKitSDK'
 
 const { trackInfo } = defineProps<{
   trackInfo: TrackInfo
 }>()
-
-const { removeVideoTrack } = useQall()
 
 const videoElement = useTemplateRef<HTMLVideoElement>('videoElement')
 const volume = ref(1)
@@ -31,54 +28,5 @@ onUnmounted(() => {
 </script>
 
 <template>
-  <div :id="'camera-' + trackInfo.participantIdentity">
-    <div>
-      <p>{{ trackInfo.participantIdentity }}</p>
-    </div>
-  </div>
-  <div :id="'camera-' + trackInfo.participantIdentity">
-    <div :class="$style.UserCard">
-      <video
-        v-if="trackInfo.trackPublication"
-        :id="trackInfo.trackPublication.trackSid"
-        ref="videoElement"
-        :class="$style.video"
-      ></video>
-      <div :class="$style.NameLabel">{{ trackInfo.participantIdentity }}</div>
-    </div>
-    <input v-model="volume" type="range" min="0" max="1" step="0.01" />
-    <button
-      v-if="!trackInfo.isRemote && trackInfo.trackPublication"
-      @click="removeVideoTrack(trackInfo.trackPublication)"
-    >
-      Remove Screen Share
-    </button>
-  </div>
+  <video v-if="trackInfo.trackPublication" ref="videoElement"></video>
 </template>
-
-<style lang="scss" module>
-.video {
-  width: 100%;
-  height: 100%;
-}
-.NameLabel {
-  position: absolute;
-  left: 8px;
-  bottom: 8px;
-  display: flex;
-  padding: 4px 12px;
-  align-items: flex-start;
-  gap: 8px;
-  border-radius: 8px;
-  background: rgba(0, 0, 0, 0.5);
-  color: #fff;
-}
-.UserCard {
-  height: 108px;
-  width: 192px;
-  // border: 1px solid black;
-  position: relative;
-  overflow: hidden;
-  border-radius: 12px;
-}
-</style>
